Move set_transform call before add_child call for consistency
Most nodes are created with PackedScene::instantiate, so it makes sense
that developers of tool scripts will adapt to its quirks. Therefore, we
should try to make the other methods of creating instances as consistent
with it as possible.

In PackedScene::instantiate, all properties except owner are set before
attaching a node to its parent. This includes transform for Node3D
instances.

Currently, Node3DEditorViewport::_create_instance sets the transform
after calling add_child. This transform is the destination used for
drag-and-drop instantiation.

A tool script that implements _ready or _enter_tree has an opportunity to
act as soon as add_child is called. If transform hasn't been set yet,
scripts can read the uninitialized property and may incorrectly assume
that the node's initial transform is equal to its default.

It's also possible for scripts to respond to
NOTIFICATION_LOCAL_TRANSFORM_CHANGED before being added to the tree.

I've moved the set_tranform call before add_child. I've also removed
the call from undo_redo. It doesn't seem necessary to redo it when we
aren't planning to undo it too.